+++ dsigm/__init__.py
@@ -1,5 +1,6 @@
 from ._dsigm import DSIGM
 from ._core import Core, CoreCluster
 from ._sgmm import SGMM
+from ._gmm import GMM
 
-__all__ = ['DSIGM', 'Core', 'CoreCluster', 'SGMM']
+__all__ = ['DSIGM', 'Core', 'CoreCluster', 'SGMM', 'GMM']
